Move messageBox.js into common folder
Auditors:

Test Plan: n/a
